directory renamed
