+++ .gitignore
@@ -19,3 +19,4 @@ target/
 
 ##
 *.log
+.env_dev_ui
