+++ src/main/java/org/voyager/app/Application.java
@@ -1,4 +1,4 @@
-package org.voyager;
+package org.voyager.app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
