+++ src/main/java/org/voyager/app/model/AirportDetails.java
