+++ src/main/java/org/voyager/app/model/AirportFilter.java
