+++ src/main/java/org/voyager/app/model/FlightDetails.java
