+++ src/main/java/org/voyager/app/model/LocationDetails.java
