+++ src/main/java/org/voyager/app/model/LocationFilter.java
