+++ src/main/java/org/voyager/app/model/ResultDetails.java
