+++ src/main/java/org/voyager/app/model/TripFilter.java
