+++ src/main/java/org/voyager/app/service/impl/AirlineServiceAPI.java
@@ -1,4 +1,4 @@
-package org.voyager.service.impl;
+package org.voyager.app.service.impl;
 
 import org.voyager.sdk.model.AirlineQuery;
 import org.voyager.commons.model.airline.Airline;
@@ -7,7 +7,7 @@ import org.voyager.sdk.service.AirlineService;
 import java.util.List;
 import java.util.Set;
 
-import static org.voyager.service.impl.VoyagerServiceImpl.unwrapEither;
+import static org.voyager.app.service.impl.VoyagerServiceImpl.unwrapEither;
 
 public class AirlineServiceAPI {
     private final AirlineService airlineService;
