+++ src/main/java/org/voyager/app/service/impl/CountryServiceAPI.java
@@ -1,4 +1,4 @@
-package org.voyager.service.impl;
+package org.voyager.app.service.impl;
 
 import lombok.NonNull;
 import org.voyager.commons.model.country.Country;
@@ -7,7 +7,7 @@ import org.voyager.sdk.service.CountryService;
 import java.util.List;
 import java.util.Optional;
 
-import static org.voyager.service.impl.VoyagerServiceImpl.unwrapEither;
+import static org.voyager.app.service.impl.VoyagerServiceImpl.unwrapEither;
 
 public class CountryServiceAPI {
     private final CountryService countryService;
