+++ src/main/java/org/voyager/app/service/impl/FlightServiceAPI.java
@@ -1,4 +1,4 @@
-package org.voyager.service.impl;
+package org.voyager.app.service.impl;
 
 import org.voyager.sdk.model.FlightQuery;
 import org.voyager.commons.model.flight.Flight;
@@ -6,7 +6,7 @@ import org.voyager.sdk.service.FlightService;
 
 import java.util.List;
 
-import static org.voyager.service.impl.VoyagerServiceImpl.unwrapEither;
+import static org.voyager.app.service.impl.VoyagerServiceImpl.unwrapEither;
 
 public class FlightServiceAPI {
     private final FlightService flightService;
