+++ src/main/java/org/voyager/app/service/impl/LocationServiceAPI.java
@@ -1,8 +1,8 @@
-package org.voyager.service.impl;
+package org.voyager.app.service.impl;
 
 import org.voyager.commons.model.location.*;
 import org.voyager.sdk.service.LocationService;
-import static org.voyager.service.impl.VoyagerServiceImpl.unwrapEither;
+import static org.voyager.app.service.impl.VoyagerServiceImpl.unwrapEither;
 
 public class LocationServiceAPI {
     private final LocationService locationService;
