+++ src/main/java/org/voyager/app/utils/ConstantsUI.java
@@ -1,4 +1,4 @@
-package org.voyager.utils;
+package org.voyager.app.utils;
 
 public class ConstantsUI {
     public static final String AIRPORT_FILTER_PARAM_NAME = "airportFilter";
